backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
anthropic>=0.40.0
openpyxl==3.1.2
rapidfuzz==3.9.3
aiofiles==23.2.1